==> Vehicle Test Data Analytics Platform/3. Data Cleaning Vehicle Log .ipynb <==
# PART 3 : Handling missing data and Quality check 

In this section, we will go through the usual data check to maintain good and useful data. The check marks will be : 
1. Check for Data Types 
2. Check for missing values (It will be divided seperately b/w EVs and Hybrid models)
3. Check for null values 
4. Check for duplicates 
 

import pandas as pd

df = pd.read_csv("combined_vehicle_test_logs.csv", parse_dates=['timestamp'])


df

# 1. Check the data types 

print(df.dtypes)


# pd.to_datetime → Turns date strings into real datetime objects for time-based analysis.

df['timestamp'] = pd.to_datetime(df['timestamp'])


# 2. Check for missing values 
# Calculates how many missing values each column has and prints only the columns where the count is greater than zero.

missing = df.isnull().sum()
print(missing[missing > 0])


# 3 . Check for null values 
# Filters out Data Frame by propulsion Type 

ev_df = df[df['propulsion_type'] == 'EV']
hybrid_df = df[df['propulsion_type'] == 'Hybrid']


# Will check if there are any missing values for EV's

print("EV missing values:\n", ev_df.isnull().sum())

# Should expect:
# fuel_rate == 0 (always) because EVs do not have fuel rate and costs per charge (This data column can be ignored )
# gear == NaN (valid as EVs don't shift gears)
# State of the charge (SOC) cannot be Null 
ev_df['soc'].isnull().sum()         # Should be 0


Here, we can see that gears have missing values Ev's do not have any gear controls.....

Here, we can see that State of Charge (Soc) is 0, as plug in hybrids use traditional fuel wit battery supply.....

df_cleaned = pd.concat([ev_df, hybrid_df], ignore_index=True)

# 4. Check for Duplicate Values 

duplicates = df.duplicated().sum()
print(f"Duplicate rows: {duplicates}")

df = df.drop_duplicates()


There no Duplicates present in the dataset....

df.to_csv("cleaned_vehicle_test_logs.csv", index=False)
